Configurable image naming.
Image names are now controlled from .evn file.

As well as which mode (dev, devssr, prod, prod-nossr, live) to run is controlled
by the .env file.

This means that it's simpler to do make commands.

Just do 'make up' and depending on variable in .env, the correct
mode is started.